renamed and moved clippath file to components, fixed require paths and clippath tests
