+++ src/modules/component/scale.js
