+++ test/unit/index.html
@@ -33,7 +33,6 @@
         "specs/modules/element/ellipse",
         "specs/modules/element/image",
         "specs/modules/element/line",
-        "specs/modules/element/clipPath",
         "specs/modules/element/path",
         "specs/modules/element/text",
         "specs/modules/helpers/add_event_listener",
@@ -43,6 +42,7 @@
         "specs/modules/layout/base",
         "specs/modules/layout/grid",
         "specs/modules/component/events",
+        "specs/modules/component/clippath",
         "specs/modules/component/brush",
         "specs/modules/component/axis",
         "specs/modules/functor",
